Split tests into smaller suites
